moved menu inflating and handling of base menu items into BaseActivity
fixes CAT-1055
