+++ catroid/res/menu/menu_current_project.xml
@@ -56,10 +56,6 @@
         android:title="@string/move"
         android:visible="false"/>
     <item
-        android:id="@+id/settings"
-        android:showAsAction="never"
-        android:title="@string/settings"/>
-    <item
         android:id="@+id/upload"
         android:showAsAction="never"
         android:title="@string/upload_button">
