+++ catroid/res/menu/menu_myprojects.xml
@@ -40,9 +40,5 @@
         android:id="@+id/show_details"
         android:showAsAction="never"
         android:title="@string/show_details"/>
-    <item
-        android:id="@+id/settings"
-        android:showAsAction="never"
-        android:title="@string/settings"/>
 
 </menu>
