+++ catroid/res/menu/menu_script_activity.xml
@@ -67,9 +67,5 @@
         android:id="@+id/show_details"
         android:showAsAction="never"
         android:title="@string/show_details"/>
-    <item
-        android:id="@+id/settings"
-        android:showAsAction="never"
-        android:title="@string/settings"/>
 
 </menu>
